+++ packages/session-replay-browser/src/config/local-config.ts
@@ -1,5 +1,5 @@
 import { Config, Logger, FetchTransport, LogLevel } from '@amplitude/analytics-core';
-import { DEFAULT_SAMPLE_RATE, DEFAULT_SERVER_ZONE } from '../constants';
+import { DEFAULT_PERFORMANCE_CONFIG, DEFAULT_SAMPLE_RATE, DEFAULT_SERVER_ZONE } from '../constants';
 import { SessionReplayOptions, StoreType } from '../typings/session-replay';
 import {
   SessionReplayLocalConfig as ISessionReplayLocalConfig,
@@ -32,6 +32,10 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
   performanceConfig?: SessionReplayPerformanceConfig;
   experimental?: { useWebWorker: boolean };
   applyBackgroundColorToBlockedElements?: boolean;
+  omitElementTags?: {
+    script?: boolean;
+    comment?: boolean;
+  };
 
   constructor(apiKey: string, options: SessionReplayOptions) {
     const defaultConfig = getDefaultConfig();
@@ -53,7 +57,7 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
     this.trackServerUrl = options.trackServerUrl;
     this.shouldInlineStylesheet = options.shouldInlineStylesheet;
     this.version = options.version;
-    this.performanceConfig = options.performanceConfig;
+    this.performanceConfig = options.performanceConfig || DEFAULT_PERFORMANCE_CONFIG;
     this.storeType = options.storeType ?? 'idb';
     this.applyBackgroundColorToBlockedElements = options.applyBackgroundColorToBlockedElements ?? false;
 
@@ -73,6 +77,9 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
     }
     if (options.experimental) {
       this.experimental = options.experimental;
+    }
+    if (options.omitElementTags) {
+      this.omitElementTags = options.omitElementTags;
     }
   }
 }