+++ packages/plugin-global-user-properties/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.65](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-global-user-properties@1.2.64...@amplitude/plugin-global-user-properties@1.2.65) (2025-06-30)
+
+**Note:** Version bump only for package @amplitude/plugin-global-user-properties
+
+
+
+
+
+## [1.2.64](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-global-user-properties@1.2.63...@amplitude/plugin-global-user-properties@1.2.64) (2025-06-26)
+
+**Note:** Version bump only for package @amplitude/plugin-global-user-properties
+
+
+
+
+
 ## [1.2.63](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-global-user-properties@1.2.62...@amplitude/plugin-global-user-properties@1.2.63) (2025-06-25)
 
 **Note:** Version bump only for package @amplitude/plugin-global-user-properties